requirements.txt
# GXII-LFEX Timing Analysis GUI - Required Python Packages
# v1.3.4 - ConfigManager統合版

# Core packages with version constraints for stability
streamlit>=1.28.0,<2.0.0
numpy>=1.21.0,<2.0.0
matplotlib>=3.5.0,<4.0.0
scipy>=1.7.0,<2.0.0
pandas>=1.3.0,<3.0.0

# Supporting packages
pyarrow>=7.0.0,<25.0.0
PyYAML>=6.0.0,<7.0.0
Pillow>=9.0.0,<11.0.0

# Note: NumPy 2.0+ causes compatibility issues with this application
# Keep NumPy < 2.0.0 for stable operation
